+++ gonggibap/src/app/page.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import Script from "next/script";
 import { Polygon, RestaurantDetailCategory } from "@/types/restaurant";
 import { ThemeToggleBtn } from "@/app/_components/ThemeToggleBtn";
-import { Sidebar } from "@/app/_components/Sidebar";
+import { Sidebar } from "@/app/_components/sidebar/Sidebar";
 import { CategoryFilter } from "@/app/_components/CategoryFilter";
 import { useMapMarkers } from "@/hooks/useMapMarkers";
 import { useKakaoMap } from "@/hooks/useKakaoMap";
